sub/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.2.1
fastapi==0.127.1
uvicorn==0.40.0
jinja2==3.1.6